+++ changelog.md
@@ -3,14 +3,14 @@ A project that using SmartFusion2 to control a gyroscope MPU6050 on the GY521 mo
 
 -----------------------------------
 ## To-Do
+- [ ] using other way to delay
 - [ ] calibrate GY521 data
 - [ ] add time in packet
 - [ ] command receive handler
 
 -----------------------------------
-## Unreleased v0.5.2(2023.07.##) @master
+## Unreleased v0.6.1(2023.07.##) @master
 ### New
-1. add delay before next uart transmission
 ### Changed
 ### Removed
 ### Deprecated
@@ -18,6 +18,13 @@ A project that using SmartFusion2 to control a gyroscope MPU6050 on the GY521 mo
 ### Known Issues
 ### Notes
 ### Reference
+
+-----------------------------------
+## BT-Gyro v0.6.0(2023.07.04) @master
+### New
+1. add delay before next uart transmission
+### Removed
+1. remove packet 5 data as one 100-bytes packet
 
 -----------------------------------
 ## BT-Gyro  v0.5.1(2023.07.04) @master